use schedule:work instead of running schedule:run every minute
This effectively does the same thing (and intermittent running
of `php artisan schedule:run` is the preferred method from the
laravel documentation), but using a loop with `sleep 60` between
invocations of schedule:run was leading to a lot of cpu throttling
of the scheduler container on openshift.

Of course, the other approach could be to run the scheduled tasks
directly from k8s CronJobs, but that's a consideration for antother
day.